+++ lib/any_date.dart
@@ -2,3 +2,5 @@
 library any_date;
 
 export 'src/any_date_base.dart' show AnyDate, DateParserInfo, Month, Weekday;
+
+export 'src/any_date_rules_model.dart' show DateParsingFunction;